Moved InitialModifyCOmmand.java to modify package.
